requirements.txt
python-telegram-bot==13.7
Pillow==9.5.0
torch==2.0.0+cpu
torchvision==0.15.0+cpu
rembg[u2net]==2.0.50
numpy==1.24.3
requests==2.31.0
